Added console methods for loading an creating and importig accounts.
* Moved account related methods to new `account.go` file
